backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9
bleach>=6.0.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
